feat: implement search functionality and layout enhancements in Dashboard
